requirements.txt
langchain
langchain_community
openai==0.28.0
huggingface_hub
streamlit
python-dotenv